reach: add reach cc api
