+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "support/iam",
     "support/auth",
     "support/enhance-wasm",
+    "support/reach",
     "services/*",
     "sdk/*",
 ]
