--- spi/spi-reach/src/domain/mod.rs
@@ -1,5 +1,0 @@
-pub mod reach_message;
-pub mod reach_message_log;
-pub mod reach_message_signature;
-pub mod reach_message_template;
-pub mod reach_trigger_global_config;
