--- spi/spi-reach/src/lib.rs
@@ -1,3 +1,0 @@
-mod dto;
-mod serv;
-mod domain;
