--- spi/spi-reach/src/serv/client/mod.rs
@@ -1,2 +1,0 @@
-mod sms;
-mod email;
