--- spi/spi-reach/src/serv/mod.rs
@@ -1,5 +1,0 @@
-pub mod client;
-
-mod message;
-mod message_signature;
-mod message_template;
