+++ support/reach/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "bios-reach"
+version.workspace = true
+authors.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+repository.workspace = true
+license.workspace = true
+edition.workspace = true
+readme.workspace = true
+publish.workspace = true
+
+[lib]
+name = "bios_reach"
+path = "src/lib.rs"
+
+[features]
+
+[dependencies]
+serde.workspace = true
+lazy_static.workspace = true
+tardis = { workspace = true, features = ["reldb-postgres", "web-server"] }
+bios-basic = { path = "../../basic", features = ["default"] }
+strum = "0.25"
+strum_macros = "0.25"
+[dev-dependencies]
+tardis = { workspace = true, features = ["test"] }
+
+bios-basic = { path = "../../basic", features = ["default", "test"] }
